chore: optimize map and restore with initial filters
