+++ packages/antd/src/hooks/table/useTable/useTable.ts
@@ -1,4 +1,4 @@
-import React, { Children, createElement, Fragment } from "react";
+import React, { Children, createElement, Fragment, useMemo } from "react";
 import { Grid, FormProps, Form, TablePaginationConfig, TableProps } from "antd";
 import { QueryObserverResult } from "@tanstack/react-query";
 import { useForm as useFormSF } from "sunflower-antd";
@@ -17,12 +17,15 @@ import {
     useTable as useTableCore,
     useTableProps as useTablePropsCore,
     useTableReturnType as useTableCoreReturnType,
+    CrudFilter,
+    ConditionalFilter,
+    LogicalFilter,
 } from "@pankod/refine-core";
 
 import {
     mapAntdSorterToCrudSorting,
     mapAntdFilterToCrudFilter,
-} from "../../../definitions/table";
+} from "@definitions/table";
 import { PaginationLink } from "./paginationLink";
 
 export type useTableProps<
@@ -120,6 +123,15 @@ export const useTable = <
         dataProviderName,
     });
 
+    const mapInitialFilter: Record<string, CrudFilter> = useMemo(() => {
+        return (initialFilter ?? []).reduce((acc, item) => {
+            const field =
+                (item as ConditionalFilter).key ||
+                (item as LogicalFilter).field;
+            return { ...acc, [field]: item };
+        }, {});
+    }, [initialFilter]);
+
     const breakpoint = Grid.useBreakpoint();
 
     const [form] = Form.useForm<TSearchVariables>();
@@ -144,6 +156,7 @@ export const useTable = <
             const crudFilters = mapAntdFilterToCrudFilter(
                 tableFilters,
                 filters,
+                mapInitialFilter,
             );
             setFilters(crudFilters);
         }
